// src/enum/grid-cell-type.enum.ts
export enum GRID_CELL_TYPE {
    CELL_TEXT = 'CELL_TEXT',
    CELL_TEXT_INFO = 'CELL_TEXT_INFO',
    CELL_TEXT_UPPERCASE = 'CELL_TEXT_UPPERCASE',
    CELL_TEXT_LOWERCASE = 'CELL_TEXT_LOWERCASE',
    CELL_TEXT_CAPITALIZE = 'CELL_TEXT_CAPITALIZE',
    CELL_BOOLEAN = 'CELL_BOOLEAN',
    CELL_DATE = 'CELL_DATE',
    CELL_NUMBER = 'CELL_NUMBER',
    CELL_LOGO = 'CELL_LOGO',
    CELL_IMAGE = 'CELL_IMAGE',
    CELL_PHONE = 'CELL_PHONE',
    CELL_PHONE_LIST = 'CELL_PHONE_LIST',
    CELL_EMAIL = 'CELL_EMAIL',
    CELL_PARENT_CATEGORY_NAME = 'CELL_PARENT_CATEGORY_NAME',
    CELL_SUB_CATEGORIES_NAME = 'CELL_SUB_CATEGORIES_NAME',
    CELL_CATEGORY_ATTRIBUTIES_NAME = 'CELL_CATEGORY_ATTRIBUTIES_NAME',
    // CELL_CATEGORY_IMAGE = 'CELL_CATEGORY_IMAGE',
    // CELL_ENTITY_NAME = 'CELL_ENTITY_NAME',
    // CELL_STORE_LOGO = 'CELL_STORE_LOGO',
    // CELL_GENDER = 'CELL_GENDER',
}